Inline all Matrix methods
This speeds up multiply_naive by more than 3x.
